+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 - Moved `for (...)` parenthesis from `for_stmt` into `for_cond`
 - Renamed `module` -> `module_declaration`
 - Renamed `var_decl` -> `var_declaration`
+- Renamed `attribute_param_list` -> `attribute_parameter_list`
+- Renamed `enum_param_list` -> `enum_parameter_list`
 
 ## 0.6.0
 - Added parsing of doc comment contracts (#33)